% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin.check_ps}
\alias{admin.check_ps}
\title{Check the integrity of practice set}
\usage{
admin.check_ps(short)
}
\arguments{
\item{short}{for the short name of the practice set}
}
\description{
Intended for teaching assistants and instructors only,
\code{admin.ps()} checks the integrity of a practice set. It loads
the source file and provides feedback on the mark-up.
}
