% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin.create_answers}
\alias{admin.create_answers}
\title{Create an .R file of answers to a practice set}
\usage{
admin.create_answers(short)
}
\arguments{
\item{short}{the unique identifier for a practice set}
}
\description{
Intended for teaching assistants and instructors only,
this function can be used to generate the "answers"
for a practice set. In other words, this function
can be used to simulate a student's work. The answers
come from the expected code, as specific in the
practice set.  Practically, this function helps you
to debug your practice set.
}
