% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin.gen_answers}
\alias{admin.gen_answers}
\title{Generates the answer files and}
\usage{
admin.gen_answers(dir_path)
}
\arguments{
\item{dir_path}{the direct into which answer files are to be saved}
}
\description{
Intended for teaching assistants and instructors only, \code{admin.gen_answers()}
creates answer files for all loaded practice sets and stores the answers in
a directory.
}
