% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin.grade}
\alias{admin.grade}
\title{Evaluate a directory of practice sets}
\usage{
admin.grade(dir = "~/Documents/_Code2/assignments/A01", short = "P01")
}
\arguments{
\item{dir}{directory of practice sets to grade}

\item{short}{for the short name of the practice set}
}
\description{
Intended for teaching assistants and instructors only, \code{admin.grade()},
will check all of the practice sets within a directory. This function will be
extended to provide auto-grader functionality.
}
