% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin.grade_repos}
\alias{admin.grade_repos}
\title{Evaluate a directory of practice sets that are structured as
follows:
/A1/user-name1/analysis.R
/A1/user-name2/analysis.R
/A1/user-name3/analysis.R}
\usage{
admin.grade_repos(filename)
}
\arguments{
\item{filename}{a directory (topmost assignment directory)}
}
\description{
This reports are placed within a sub-folder, named \code{results}.
}
