% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin.grade_ui}
\alias{admin.grade_ui}
\title{UI for selecting files to grade}
\usage{
admin.grade_ui(short = "P01")
}
\arguments{
\item{short}{for the short name of the practice set}
}
\description{
UI for selecting files to grade
}
