% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin.grade_ui_file}
\alias{admin.grade_ui_file}
\title{Grade a single file}
\usage{
admin.grade_ui_file()
}
\description{
Intended for teaching assistants and instructors only, this function allows
you to select a single file and grade it.
}
