% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin.load}
\alias{admin.load}
\title{Load practice sets into the package}
\usage{
admin.load(dir)
}
\arguments{
\item{dir}{a directory}
}
\description{
Intended for teaching assistants and instructors only, \code{admin.load()}
is used to load practice sets. All practice sets that are found in the
the directory, \code{dir}, are loaded into the package. (Default practice
sets are automatically added.  These defaults are installed when the
package is installed.)
}
