% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin.ls}
\alias{admin.ls}
\title{List installed practice sets}
\usage{
admin.ls()
}
\description{
Intended for instructors, \code{admin.ls()}
simply lists all currently installed practice sets. Each of these
practice sets are available to learners and for grading.
}
\details{
To grade student work, the practice set must be loaded.
}
