% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin}
\alias{admin}
\title{List the available admin functions}
\usage{
admin()
}
\description{
Intended for teaching assistants and instructors only,
\code{admin} simply lists all the available admin functions.
}
