% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin.vars}
\alias{admin.vars}
\title{List practice set objects}
\usage{
admin.vars()
}
\description{
Intended for teaching assistants and instructors only,
\code{admin.prompts()} lists all of the objects associated with
the currently active practice set, including expected variables
and callback functions for checking code.
}
