% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_practice_sets.R
\name{create_practice_set}
\alias{create_practice_set}
\title{Create practice set from a coded text file}
\usage{
create_practice_set(ps, fn)
}
\arguments{
\item{ps}{a practice set}
}
\value{
a filename
}
\description{
Xxx xxx xxx
}
