% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps.R
\name{DEFAULT_Check}
\alias{DEFAULT_Check}
\title{Default function for checking learner's code}
\usage{
DEFAULT_Check(internal_id, result)
}
\arguments{
\item{internal_id}{the internal ID of the prompt from the practice set}

\item{result}{the result, which grows upon each call to a <var>_Check function}

\item{learner_val}{the value of the prompt variable}
}
\description{
Default function for checking learner's code
}
