% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps.R
\name{eval_expr}
\alias{eval_expr}
\title{Evaluate an R expression}
\usage{
eval_expr(id)
}
\arguments{
\item{id}{the internal ID for the prompt}
}
\value{
returns the result of the evaluation. If the learner's code produces
a function (that is, a "closure") a text string summarizing the function is
returned
}
\description{
To test a learner's answer, call this function. This function accesses
the learner's code, which is stored in the practice set data structure.
}
