% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice-info201.R
\docType{data}
\name{pkg.globals}
\alias{pkg.globals}
\title{#' @import miniUI
NULL}
\format{
An object of class \code{environment} of length 3.
}
\usage{
pkg.globals
}
\description{
#' @import stringr
NULL
}
\details{
#' @import shiny
NULL

#' @import rstudioapi
NULL
}
\keyword{datasets}
