% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps.R
\name{practice.answers}
\alias{practice.answers}
\title{Show all the answers for the practice set}
\usage{
practice.answers()
}
\value{
`TRUE` if all goes well.
}
\description{
Show all the answers for the practice set
}
\seealso{
\code{\link{practice.questions}}
}
