% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student.R
\name{practice.begin}
\alias{practice.begin}
\title{Begin a practice set}
\usage{
practice.begin(short = "P01", learner = "Anonymous", email = "")
}
\arguments{
\item{short}{The short ID for this practice set.}

\item{learner}{The learner's name}

\item{email}{The learner's e-mail address}
}
\value{
`TRUE` if all goes well; otherwise, this function will stop with an message.
}
\description{
Call this function to install a specific practice
set and to set things up. Currently, there are two practice sets -
"P01" and "P02".
}
