% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps.R
\name{practice.check}
\alias{practice.check}
\title{Check the practice set answers.}
\usage{
practice.check()
}
\value{
`TRUE` if all goes well; otherwise, this function will stop with an message.
}
\description{
Check the practice set answers.
}
\seealso{
\code{\link{practice.questions}}
}
