% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps.R
\name{practice.load_url}
\alias{practice.load_url}
\title{Read a practice set}
\usage{
practice.load_url(
  fn = paste0("https://raw.githubusercontent.com/",
    "info201B-2022-Autumn/practice-sets/main/", "PS-T10.R")
)
}
\arguments{
\item{fn}{filename}
}
\description{
Read a practice set
}
