% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice-info201.R
\name{practice.questions}
\alias{practice.questions}
\title{Show the questions for a practice set.}
\usage{
practice.questions()
}
\value{
`TRUE` if all goes well.
}
\description{
Show the questions for a practice set.
}
\examples{
# Show practice questions for default practice set
practice.begin()
practice.questions()

# Show practice questions for practice set "P02"
practice.begin("P02")
practice.questions()
}
\seealso{
[practice.questions]
}
