% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner.R
\name{practice.setwd}
\alias{practice.setwd}
\title{Set a path to a root directory for R practice sets}
\usage{
practice.setwd(dir_path)
}
\arguments{
\item{dir_path}{An absolute path directory working directory}
}
\description{
Set a path to a root directory for R practice sets
}
