% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps.R
\name{result_error_msg}
\alias{result_error_msg}
\title{Formats a "try again" message}
\usage{
result_error_msg(id, show_hints = TRUE)
}
\arguments{
\item{id}{the internal ID for the prompt}

\item{show_hints}{to show or not show the hints for incorrect answers}
}
\value{
a formatted string that can be concatenated to a full output string
}
\description{
This should be called when a learner's answer is incorrect.
}
