% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice-info201.R
\name{result_good_msg}
\alias{result_good_msg}
\title{Formats a "good" message}
\usage{
result_good_msg(id)
}
\arguments{
\item{id}{the internal ID for the prompt}
}
\value{
a formatted string that can be concatenated to a full output string
}
\description{
This should be called when a learner's answer is correct.
}
