% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice-info201.R
\name{result_main_message}
\alias{result_main_message}
\title{Formats a message}
\usage{
result_main_message(main_message)
}
\arguments{
\item{main_message}{the text of the message}
}
\value{
a formatted string that can be concatenated to a full output string
}
\description{
Currently, no formatting is done.
}
