% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice-info201.R
\name{result_msg}
\alias{result_msg}
\title{#' Xxx xxx}
\usage{
result_msg(id, is_correct, show_hints = TRUE)
}
\arguments{
\item{id}{Internal prompt ID}

\item{is_correct}{boolean indicating if the learner's answer is correct}

\item{show_hints}{to show or not show the hints}
}
\value{
xxx
}
\description{
#' Xxx xxx
}
