% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice-info201.R
\name{result_sub_message}
\alias{result_sub_message}
\title{#' Xxx xxx}
\usage{
result_sub_message(message, sub_message)
}
\arguments{
\item{message}{the current message}

\item{sub_message}{the sub-message to be added}
}
\value{
xxx
}
\description{
#' Xxx xxx
}
