% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin.grade}
\alias{admin.grade}
\title{Grade all files in a folder}
\usage{
admin.grade()
}
\description{
Intended for teaching assistants and instructors only, this function allows
you to select a folder and grade all files within this folder.
}
