% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin.grade}
\alias{admin.grade}
\title{Evaluate a directory of practice sets}
\usage{
admin.grade(filename)
}
\arguments{
\item{filename}{either a directory for a particular file}
}
\description{
Intended for teaching assistants and instructors only, \code{admin.grade()},
this function will grade all of the practice sets within a directory.
That is, it will check student code against expected code and produce
a reports - currently, one report for each student assignment.
}
\details{
This reports are placed within a sub-folder, named \code{results}.
}
