% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin.grade}
\alias{admin.grade}
\title{Evaluate a directory of practice sets}
\usage{
admin.grade(filename)
}
\arguments{
\item{filename}{either a directory for a particular file}
}
\description{
Intended for teaching assistants and instructors only, \code{admin.grade()},
will check all of the practice sets within a directory.
}
