% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin.grade_repos}
\alias{admin.grade_repos}
\title{Evaluate a directory of GitHub repos}
\usage{
admin.grade_repos(filename)
}
\arguments{
\item{filename}{a directory (topmost assignment directory)}
}
\description{
This function is similar to \code{admin.grade_fn}. The difference
is that the directory structure is assumed to be the following:

/A1/user-name1/analysis.R \cr
/A1/user-name2/analysis.R \cr
/A1/user-name3/analysis.R \cr
...

The evaluation reports are place in a sub-folder, named \code{results}.
}
