% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin.ls}
\alias{admin.ls}
\title{List installed practice sets}
\usage{
admin.ls()
}
\description{
Intended for teaching assistants and instructors only,
\code{admin.ls()} simply lists all currently installed practice sets.
At present, each of these practice sets will be available to
learners.
}
