% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{admin.prompts}
\alias{admin.prompts}
\title{Show the practice set prompts}
\usage{
admin.prompts(short, show_answers = FALSE)
}
\arguments{
\item{short}{for the short name of the practice set}
}
\description{
Intended for teaching assistants and instructors only, \code{admin.prompts()}
shows the prompts and expected answers in a practice set. The output is
R code showing the correct answers.
}
