% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice-info201.R
\name{practice.begin}
\alias{practice.begin}
\title{Begin a practice set. Call this function to install a specific practice
set and to set things up. Currently, there are two practice sets -
"P01" and "P02".}
\usage{
practice.begin(short = "P01", to_console = FALSE)
}
\arguments{
\item{short}{The short ID for this practice set.}

\item{console}{Indicates whether output should go to the console.}
}
\value{
`TRUE` if all goes well; otherwise, this function will stop with an message.
}
\description{
Begin a practice set. Call this function to install a specific practice
set and to set things up. Currently, there are two practice sets -
"P01" and "P02".
}
\examples{
practice.begin()
practice.begin("P02")
practice.begin("P02",TRUE)
}
\seealso{
\code{\link{practice.questions}}
}
