% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps.R
\name{practice.begin}
\alias{practice.begin}
\title{Begin a practice set. Call this function to install a specific practice
set and to set things up. Currently, there are two practice sets -
"P01" and "P02".}
\usage{
practice.begin(short = "P01", learner = "Anonymous")
}
\arguments{
\item{short}{The short ID for this practice set.}

\item{learner}{The learner's name}
}
\value{
`TRUE` if all goes well; otherwise, this function will stop with an message.
}
\description{
Begin a practice set. Call this function to install a specific practice
set and to set things up. Currently, there are two practice sets -
"P01" and "P02".
}
\seealso{
\code{\link{practice.questions}}
}
