% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student.R
\name{practice.check}
\alias{practice.check}
\title{Evaluate a practice set and provide feedback}
\usage{
practice.check()
}
\value{
`TRUE` if all goes well; otherwise, this function will stop with an message.
}
\description{
Evaluate a practice set and provide feedback
}
