% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student.R
\name{practice.load_url}
\alias{practice.load_url}
\title{Read and load a practice set into RStudio}
\usage{
practice.load_url(
  fn = paste0("https://raw.githubusercontent.com/",
    "info201B-2022-Autumn/practice-sets/main/", "PS-T10.R")
)
}
\arguments{
\item{fn}{The url for the practice set}
}
\description{
Read and load a practice set into RStudio
}
