% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student.R
\name{practice.questions}
\alias{practice.questions}
\title{Show questions for a practice set}
\usage{
practice.questions(do_not_show = NULL)
}
\arguments{
\item{do_not_show}{optional list of internal IDs of questions to not show}
}
\value{
\code{TRUE} if all goes well.
}
\description{
Show questions for a practice set
}
