% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice-info201.R
\name{ps_get_expected_answer}
\alias{ps_get_expected_answer}
\title{Get the expected answer}
\usage{
ps_get_expected_answer(id)
}
\arguments{
\item{id}{the internal ID for the prompt}
}
\value{
formatted code
}
\description{
Xxx xxx xxx
}
