% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice-info201.R
\name{result_error_msg}
\alias{result_error_msg}
\title{#' Xxx xxx}
\usage{
result_error_msg(id, show_hints = TRUE)
}
\arguments{
\item{id}{Internal prompt ID}

\item{show_hints}{to show or not show the hints}
}
\value{
xxx
}
\description{
#' Xxx xxx
}
