% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice-info201.R
\name{result_main_message}
\alias{result_main_message}
\title{#' Xxx xxx}
\usage{
result_main_message(main_message)
}
\arguments{
\item{main_message}{a main message}
}
\value{
xxx
}
\description{
#' Xxx xxx
}
