% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice-info201.R
\name{result_sub_message}
\alias{result_sub_message}
\title{Format a sub-message}
\usage{
result_sub_message(message, sub_message)
}
\arguments{
\item{message}{the formatted text of the current message}

\item{sub_message}{the sub-message to be added}
}
\value{
a formatted string that can be concatenated to a full output string
}
\description{
Currently, sub-messages are simply intended
}
