% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice-info201.R
\name{result_update}
\alias{result_update}
\title{#' Xxx xxx}
\usage{
result_update(result, id, is_correct, text)
}
\arguments{
\item{result}{a list of the results}

\item{id}{internal id of the prompt}

\item{is_correct}{indicating whether the answer is correct or not}

\item{text}{a message}
}
\value{
xxx
}
\description{
#' Xxx xxx
}
